code/requirements.txt
langchain
langchain-openai
langchain-chroma
openai
pandas
tqdm
python-dotenv
pydantic
tiktoken
